determine latest generation in an experiment

+++ evolver/src/main.py
@@ -8,21 +8,23 @@ import os
 ACTIONS_NUM_FOR_GRZIB = 133 # 135 - 2 bcs skipping MR and FO
 
 def latest_generation(path):
-    generation = 8
+    files = os.listdir(path)
+    file = next(reversed(files))
+    while not file.endswith(".json"):
+        file = next(files)
+    generation = int(file.partition(".json")[0].partition("generation")[2])
     r = f"{path}/generation{generation}.json"
-    print("I'm a lazy fuck, hardcoded latest")
     return r, generation
 
 def refresh_directory(path):
-    try:
+    if os.path.exists(path):
         files = os.listdir(path)
-    except FileNotFoundError:
+        for file in files:
+            file_path = os.path.join(path, file)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+    else:
         os.mkdir(path)
-        return
-    for file in files:
-        file_path = os.path.join(path, file)
-        if os.path.isfile(file_path):
-            os.remove(file_path)
 
 @click.command()
 @click.argument("experiment_name")
@@ -38,14 +40,13 @@ def train(experiment_name, fresh, seed):
 
     if fresh:
         refresh_directory(experiment_path)
-        population = genetic.fresh_population(size=100, genom_size=ACTIONS_NUM_FOR_GRZIB)
+        population = genetic.fresh_population(size=50, genom_size=ACTIONS_NUM_FOR_GRZIB)
     else:
-        population_file, generation = latest_generation(f"{experiment_path}")
-        try:
-            population = genetic.load_population(population_file)
-        except FileNotFoundError:
-            print(f"Experiment '{experiment_name}' does not exist so can't be continued. (Forgot -f?)")
+        if not os.path.exists(experiment_path):
+            print(f"Experiment '{experiment_name}' does not exist. Maybe you meant to run with -f?")
             exit(1)
+        population_file, generation = latest_generation(f"{experiment_path}")
+        population = genetic.load_population(population_file)
 
     solutions = genetic.solve(experiment_path, tb, generations_num=1000, population=population, starting_generation=generation)
     top3 = tools.selBest(solutions, k=3)
